+++ app/src/main/java/com/gem/weather/rest/dto/WeatherCountryDTO.java
@@ -38,7 +38,7 @@ public class WeatherCountryDTO extends BaseResponseDTO implements Serializable{
   private SysDTO sys;
   @SerializedName("id")
   @Expose
-  private Integer id;
+  private Long id;
   @SerializedName("name")
   @Expose
   private String name;
@@ -160,15 +160,15 @@ public class WeatherCountryDTO extends BaseResponseDTO implements Serializable{
     return this;
   }
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
-  public WeatherCountryDTO withId(Integer id) {
+  public WeatherCountryDTO withId(Long id) {
     this.id = id;
     return this;
   }
